3_Langchain_Tools_Calling/requirements.txt
langchain_openai
langchain_core
langchain
python-dotenv
